Address CR comments.
