+++ allennlp/models/__init__.py
@@ -13,4 +13,3 @@ from allennlp.models.semantic_role_labeler import SemanticRoleLabeler
 from allennlp.models.simple_tagger import SimpleTagger
 from allennlp.models.coreference_resolution.coref import CoreferenceResolver
 from allennlp.models.constituency_parser import SpanConstituencyParser
-from allennlp.models.ensemble import BidafEnsemble
