--- allennlp/models/ensemble/__init__.py
@@ -1,2 +1,0 @@
-from allennlp.models.ensemble.bidaf_ensemble import BidafEnsemble
-from allennlp.models.ensemble.bidaf_ensemble import Ensemble
