+++ allennlp/models/reading_comprehension/__init__.py
@@ -6,3 +6,4 @@ These submodules contain models for things that are predominantly focused on rea
 """
 
 from allennlp.models.reading_comprehension.bidaf import BidirectionalAttentionFlow
+from allennlp.models.reading_comprehension.bidaf_ensemble import BidafEnsemble
